Add beneficiaries field types
